pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "winterm2"
version = "0.1.0"
description = "Windows-native CLI tool for automating Windows Terminal"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "psutil>=5.9",
    "pyyaml>=6.0",
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.scripts]
wt2 = "wt2.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/winterm2"
Repository = "https://github.com/yourusername/winterm2"
Issues = "https://github.com/yourusername/winterm2/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["wt2"]
omit = ["tests/*"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

[tool.black]
line-length = 88
target-version = ["py310"]
include = "\\.pyi?$"
extend-exclude = "tests/"

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "W", "I", "UP", "B", "C4", "ARG", "SIM", "PERF"]
ignore = ["E501", "B008", "C901"]
extend-exclude = ["tests/*", ".git"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
